(BOLT-985) Add plan executor to bolt-server package
This adds the plan-executor application and service to the `pe-bolt-server` package. Since this package is expected to only be run and managed in PE, and these services are closely related (though not technically dependent on each other), it seemed prudent to just have one pipeline and package to build and ship them together. This PR adds plan-executor settings to the bolt-server project, adds a plan-executor component, and plan-executor service files.