(BOLT-1076) Bump orchestrator_client gem
Pull in orchestrator_client gem that contains enhancement whereby an error with a helpful message is raised when an invalid character is detected in a token file or if the token-file is unreadable. The new version also requires two new (to Bolt) gems that allow persistent http connections when using linux Bolt controller.