(BOLT-1220) Enable pe-bolt-server el 8 builds
Puppet agent is only shipping to http://enterprise.delivery.puppetlabs.net/2019.1/ which is required to build bolt-server on redhat 8. This commit bumps the peversion to pull packages from the 2019.1 enterprise repo. The project config for pe-bolt-server is updated to take advantage of the fact that vanagon 0.15.22 no longer enables shebang mangling. The project config also contains a special case to handle disabling build-id generation that apparently are caused by system libgcc and listdc++ on redhat 8.