(maint) Do not try to use latest system packages for el8
After moving from the beta Redhat8 to the official released version some of the pre-canned versions of core packages (like rpm and gcc) are to 'new' to be used with additional packages installed to build RPMs. This commit adds the --nobest cli option to not force the package manager to use the newest packages.