(GH-1172) Make bolt-inventory-pdb executable out of the box
Closes puppetlabs/bolt#1172

This adds a wrapper script for the `bolt-inventory-pdb` command similar
to the one we use for `bolt`. The script gets installed to
`/opt/puppetlabs/bin` and symlinked from `/usr/local/bin` which is
typically on the users path. This then calls the actual ruby script at
`/opt/puppetlabs/bolt/bin` to execute the bolt-inventory-pdb command.