(packaging) Set Docker container locale to UTF-8
In some cases, the remote side of a Bolt run will render UTF-8
characters (for example, in apply_prep). Bolt's logging mechanism uses
the the system locale, which defaults to POSIX, and will throw an
exception if it encounters UTF-8. This change sets the locale of our
container to UTF-8 globally, which remedies the problem.