(GH-1895) Add action to generate powershell module
In order to ship the powershell cmdlets module in the Bolt MSI package
we need generate the module from it's constituent files and commit it to
bolt-vanagon. A repository-dispatch event will be created and sent to
bolt-vanagon whenever the pwsh files are updated in Bolt. This commit
adds an action that will be triggered when the repository-dispatch is
received which:
* Clones both Bolt and bolt-vanagon repos
* Generates the module from the rake task in Bolt
* Moves the generated module to bolt-vanagon
* Commits and pushes it to master

This ensures that the generated module will stay up to date with changes
in Bolt.