(GH-2078) Add topic guides to PowerShell module
This adds Bolt's [topic
guides](https://github.com/puppetlabs/bolt/pull/2088) to the PowerShell
module. It updates the Bolt component to copy the renamed topic guides
from `pwsh_module/en-US` in the Bolt repo to the PowerShell module
directory. It also updates a couple of WiX files to create a directory
in the PowerShell module for the help pages and adds each of the help
pages as a component.

Part of puppetlabs/bolt#2078