(GH-146) Use Tarball method for MSI based packages
Bolt ships several gems in order to perform various functions, which contain ~20,000 files. This causes long MSI install times as MSI hashes and keeps track of each file in order to determine if it needs to be installed or not. This is a costly examination when dealing with thousands of files, but is useful in most upgrade scenarios. However, it is only useful when we 'know' the file needs to be upgraded/replaced/removed, which we only really know if we control those files. The majority of files bolt ships are external gems, which we don't control and we don't keep track of. So in short, we are paying a cost to track files we don't need to track. This commit reduces this cost by archiving the majority of files we don't track into one archive that is expanded on install time.

Copies work done in https://github.com/puppetlabs/pdk-vanagon/pull/211

This commit provides:

- Adds a new component called `bolt-create-ruby-tarballs` which is responsible for taking Bolt installation files and compressing the gem folders. This is only for Windows builds at the moment
- Adds a new file the to installation media called `extract_all.rb`. This small script is responsible for extracting the tarballs back to their original locations.
- Adds a VBScript Binary component to both extract and remove the tarball contents
- Adds a custom action to capture the INSTALLDIR as this property is not available to Deferred Custom Actions
- Adds a Deferred Custom Action to extract the tarballs so that it can be run in an elevated context
- Adds a Deferred Custom Action to remove the extracted files so that it can be run in an elevated context

Speed increase comes with new installs. Upgrades will still be slow until this version is the version being upgraded.