(maint) Migrate Bolt to bolt.bat
Migrates Bolt from the `bolt` PowerShell function to a `bolt.bat` file and adds a WiX component that appends the path to Bolt's bin directory to the system PATH Environment Variable.

This enables Bolt to be used in any shell on a Windows Platform, even Cygwin and Git Bash, not just PowerShell.

Requires work done in https://github.com/puppetlabs/bolt/pull/2532 to be complete.