(GH-2291) Manage ProgramData folder for Bolt
As part of using Bootsnap for caching in Bolt, Bootsnap needs to have a
directory to write cache to. The right directory for this is
`%PROGRAMDATA%\PuppetLabs\Bolt`, however that directory may not exist or
may not be writeable by the user installing Bolt. This updates the
Bootsnap setup script to create the directory if it doesn't already
exist, and verify that it's writeable before starting Bootsnap. If
either of these fails the script will log messages detailing what
happened, and will fail the Bolt install.